Module OpamCached

module type ARG = sig ... end

Argument type for the Make functor

module Make (X : ARG) : sig ... end

Module handling a simple marshalled cache for a given data structure. A magic number is added to validate the version of the cache.