Module type Config.Sig

type t

Read-only record type containing the lib's configuration options

type 'a options_fun

Type of functions with optional arguments for setting each of t's fields, similarly named, and returning 'a

val default : t

The default values of the options to use at startup

val set : t -> (unit -> t) options_fun

Use to update any option in a t, using the optional arguments of options_fun. E.g. set opts ?option1:1 ?option4:"x" ()

val setk : (t -> 'a) -> t -> 'a options_fun

Same as set, but passes the result to a continuation, allowing argument stacking

val r : t ref

The global reference containing the currently set library options. Access using OpamXxxConfig.(!r.field).

val update : ?noop:_ -> (unit -> unit) options_fun

Updates the currently set options in r according to the optional arguments

val init : ?noop:_ -> (unit -> unit) options_fun

Sets the options, reading the environment to get default values when unspecified

val initk : 'a -> 'a options_fun

Sets the options like init, but returns the given value (for arguments stacking)