Module OpamStd.Win32

Windows-specific functions

module RegistryHive : sig ... end

Win32 Registry Hives and Values

val set_parent_pid : int32 -> unit

Change which the pid written to by parent_putenv. This function cannot be called after parent_putenv.

val parent_putenv : string -> string -> bool

Update an environment variable in the parent (i.e. shell) process's environment.

val persistHomeDirectory : string -> unit

persistHomeDirectory value sets the HOME environment variable in this and the parent process and also persists the setting to the user's registry and broadcasts the change to other processes.