Module OpamVersionCompare

Version comparison function used throughout. From the Dose suite.

Functions for manipulating and comparing Debian version strings. Compliant with Debian policy version 3.9.2. and Debian developers reference version 3.4.6

Comparing debian version strings

The following functions compare any two strings, that is these functions do not check whether the arguments are really legal debian versions. If the arguments are debian version strings, then the result is as required by debian policy. Note that two strings may be equivalent, that is denote the same debian version, even when they differ in syntax, as for instance "0:1.2.00" and "1.02-0".

val equal : string -> string -> bool

returns
true iff the two strings define the same version. Hence, the result may be true even when the two string differ syntactically.

val compare : string -> string -> int

compare x y returns 0 if x is eqivalent to y, -1 if x is smaller than y, and 1 if x is greater than y. This is consistent with Stdlib.compare.