Module OpamCached.Make

Module handling a simple marshalled cache for a given data structure. A magic number is added to validate the version of the cache.

Parameters

module X : ARG

Signature

type t = X.t

val save : OpamFilename.t -> t -> unit

Marshal and write the cache to disk

val load : OpamFilename.t -> t option

Load the cache if it exists and is valid and compatible with the current binary. Clear it otherwise.

val remove : OpamFilename.t -> unit