Module OpamStd

Generic stdlib functions (String, List, Option, Sys submodules...)

Signatures and functors

module type SET = sig ... end

Sets with extended interface and infix operators

module type MAP = sig ... end

Maps with extended interface

module type ABSTRACT = sig ... end

A signature for handling abstract keys and collections thereof

module AbstractString : ABSTRACT with type t = string

A basic implementation of ABSTRACT using strings

Generators for set and map modules with printers

module type OrderedType = sig ... end

module Set : sig ... end

module Map : sig ... end

Integer collections

module IntMap : MAP with type key = int

Map of ints

module IntSet : SET with type elt = int

Set of ints

Utility modules extending the standard library on base types

module Option : sig ... end

module List : sig ... end

module String : sig ... end

module Format : sig ... end

module Exn : sig ... end

Manipulation and query of environment variables

module Env : sig ... end

System query and exit handling

module Sys : sig ... end

module Win32 : sig ... end

General use infix function combinators

module Op : sig ... end

Helper functions to initialise configuration from the environment

module Config : sig ... end