Module OpamStd.Env

val reset_value : prefix:string -> char -> string -> string list

Remove from a c-separated list of string the ones with the given prefix

val cut_value : prefix:string -> char -> string -> string list * string list

split a c-separated list of string in two according to the first occurrences of the string with the given prefix. The list of elements occurring before is returned in reverse order. If there are other elements with the same prefix they are kept in the second list.

val get : string -> string

val getopt : string -> string option

val list : unit -> (string * string) list

val escape_single_quotes : ?using_backslashes:bool -> string -> string

Utility function for shell single-quoted strings. In most shells, backslash escapes are not allowed and a single quote needs to be replaced by quote double-quote quote double-quote quote (close the single-quoted literal, put the single quote in a double-quoted literal, and reopen a single-quoted literal). fish is the exception and should set using_backslashes to escape both quotes and backslashes using backslashes